pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]
[project]
authors = [{name = "EpicStuff", email = "EpicStuff@users.noreply.github.com"}]
classifiers = ["Operating System :: OS Independent", "Programming Language :: Python :: 3"]
dependencies = ["epicstuff", "ruamel.yaml==0.18.10"]
description = "YAML but with tabs"
name = "taml"
readme = "README.md"
requires-python = ">=3.10"
version = "1.1.0"

[project.urls]
Homepage = "https://github.com/EpicStuff/taml"
Issues = "https://github.com/EpicStuff/taml/issues"
